Plotting TwoBody and NBody Systems

TwoBodyPropagationResult and MultibodyPropagationResult structs can be plotted with the plot function. Currently, two-body orbits can be plotted in the Perifocal (2D) frame, and the Cartesian (3D) frame. N-body systems can currently only be plotted in 3D.

UnitfulAstrodynamics.AstroPlots.orbitplot — Function

orbitplot(sols)
orbitplot(sols, frame; kwargs...)

Plots every timestep in sols in 3D space. All keyward arguments are passed directly to Plots.jl.

orbitplot(sols; bodies, kwargs...)

Plots every timestep in sols in 3D space. All keyward arguments are passed directly to Plots.jl.

Missing docstring.

Missing docstring for lagrangeplot. Check Documenter's build log for details.