Maneuvers Calculations

Currently, Maneuvers focuses on the twobody problem. The following functions have been developed. More to come!

UnitfulAstrodynamics.Maneuvers.escape_radius — Function

escape_radius(r₀, v₀, aₜ)

Provides the radius of escape.

UnitfulAstrodynamics.Maneuvers.escape_velocity — Function

escape_velocity(r₀, v₀, aₜ, μ)

Provides the velocity at escape.

UnitfulAstrodynamics.Maneuvers.escape_time — Function

escape_time(r₀, v₀, aₜ)

Provides time delta from the provided initial orbit to escape.

UnitfulAstrodynamics.Maneuvers.escape_path_length — Function

escape_path_length(r₀, v₀, aₜ)

Provides the path length from the initial condition to escape.