NBody Data Structures

As with TwoBody Data Structures, the NBody submodule includes data structures for storing multibody orbital states. The Body structure holds position, velocity, and mass information for a single body. A MultibodySystem contains an array of Body structures, and is used to completely describe a multibody orbital state.

UnitfulAstrodynamics.NBody.Body — Type

Stores the state of each body in the NBody problem.

UnitfulAstrodynamics.NBody.MultibodySystem — Type

Describes a system of n MultibodyStates's.