ThreeBody Calculations

Common ThreeBody problem calculations are provided through functions.

Frame Representations

You can convert between the Inertial and Rotating (Synodic) reference frames through the inertial and synodic functions.

UnitfulAstrodynamics.ThreeBody.inertial — Function

inertial(vecₛ, t)
inertial(vecₛ, t, ω)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:203.

Given the Synodic frame vector, returns the vector in the inertial reference frame.

UnitfulAstrodynamics.ThreeBody.synodic — Function

synodic(rᵢ, vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:217.

Returns the position and velocity vectors in the synodic (rotating) reference frame.

Dimensionalization

Functions to nondimensionalize spacecraft states, and re-dimensionalize spacecraft states are provided.

UnitfulAstrodynamics.ThreeBody.time_scale_factor — Function

time_scale_factor(a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:10.

Returns time scale factor, Tₛ.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_length — Function

nondimensionalize_length(rᵢ, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:15.

Returns nondimensional length unit, DU.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_velocity — Function

nondimensionalize_velocity(vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:20.

Returns nondimensional velocity unit, DU/DT.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_time — Function

nondimensionalize_time(t, a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:25.

Returns nondimensional time unit, DT.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_mass_parameter — Function

nondimensionalize_mass_parameter(μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:30.

Returns nondimensional mass parameter, μ.

UnitfulAstrodynamics.ThreeBody.nondimensionalize — Function

nondimensionalize(rᵢ, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:37.

Returns nondimensional form of (Unitful) scalar posiion.

nondimensionalize(rᵢ, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:44.

Returns nondimensional form of (Unitful) position vector.

nondimensionalize(vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:51.

Returns nondimensional form of (Unitful) scalar velocity.

nondimensionalize(vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:58.

Returns nondimensional form of (Unitful) velocity vector.

nondimensionalize(vᵢ, a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:65.

Returns nondimensional form of (Unitful) velocity vector.

nondimensionalize(t, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:72.

Returns nondimensional form of (Unitful) time duration.

nondimensionalize(t, a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:79.

Returns nondimensional form of (Unitful) time duration.

nondimensionalize(μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:86.

Returns nondimensional form of (Unitful) graviational parameters.

nondimensionalize(r₃, v₃, Δt, μ₁, μ₂, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:98.

Returns nondimensional Circular Restricted Three-body State.

nondimensionalize(state)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:110.

Returns the nondimensional (synodic / rotating) representation of a CR3BP state.

UnitfulAstrodynamics.ThreeBody.redimensionalize_length — Function

redimensionalize_length(rᵢ, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:122.

Returns dimensional length units.

UnitfulAstrodynamics.ThreeBody.redimensionalize_velocity — Function

redimensionalize_velocity(vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:127.

Returns dimensional velocity units.

UnitfulAstrodynamics.ThreeBody.redimensionalize_time — Function

redimensionalize_time(t, a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:132.

Returns dimensional time unit.

redimensionalize_time(t, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:137.

Returns dimensional time unit.

UnitfulAstrodynamics.ThreeBody.redimensionalize — Function

redimensionalize(rᵢ, a)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:144.

Returns dimensional (inertial) form of (Unitful) scalar posiion.

redimensionalize(vᵢ, a, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:151.

Returns dimensional (inertial) form of (Unitful) velocity vector.

redimensionalize(t, Tₛ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:158.

Returns dimensional (inertial) form of (Unitful) time duration.

redimensionalize(t, a, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:165.

Returns dimensional (inertial) form of (Unitful) time duration.

redimensionalize(state, μ₁, μ₂)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:174.

Returns the dimensional (inertial) representation of a CR3BP state.

Halo Orbit Solvers

UnitfulAstrodynamics.ThreeBody.analyticalhalo — Function

analyticalhalo(μ; Az, ϕ, steps, L, hemisphere)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:68.

Returns an analytical solution for a Halo orbit about L.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

Az: Desired non-dimensional Z-amplitude for Halo orbit.

ϕ: Desired Halo orbit phase.

steps: Number of non-dimensional timepoints in returned state.

L: Lagrange point to orbit (L1 or L2).

hemisphere: Specifies northern or southern Halo orbit.

Outputs:

Synodic position vector r::Array{<:AbstractFloat}

Synodic velocity vector v::Array{<:Abstractfloat}.

Halo orbit period Τ.

Throws ArgumentError if L is not 1 or 2.

References:

Rund, 2018.

UnitfulAstrodynamics.ThreeBody.halo — Function

halo(μ; Az, L, hemisphere, tolerance, max_iter, reltol, abstol)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:172.

Returns a numerical solution for a Halo orbit about L.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

Az: Desired non-dimensional Z-amplitude for Halo orbit.

ϕ: Desired Halo orbit phase.

L: Lagrange point to orbit (L1 or L2).

hemisphere: Specifies northern or southern Halo orbit.

Outputs:

Tuple of initial states: (r, v) where r::Vector{<:AbstractFloat}, v::Vector{<:Abstractfloat}.

Throws ArgumentError if L is not :L1 or :L2

References:

Rund, 2018.

halo(μ; Az, L, hemisphere, tolerance, max_iter, reltol, abstol)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:172.

halo(μ, DU, DT; kwargs...)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:250.

Returns a NondimensionalThreeBodyState type, instead of tuple r,v,T.

UnitfulAstrodynamics.ThreeBody.monodromy — Function

monodromy(orbit; check_periodicity, reltol, abstol, atol)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:299.

Returns the Monodromy Matrix for a Halo orbit.

Other Common Calculations

UnitfulAstrodynamics.ThreeBody.potential_energy — Function

potential_energy(r, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:191.

Returns the potential energy U.

UnitfulAstrodynamics.ThreeBody.jacobi_constant — Function

jacobi_constant(r, v, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:196.

Returns the Jacobi Constant C.

UnitfulAstrodynamics.ThreeBody.lagrange — Function

lagrange(μ)
lagrange(μ, L)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:235.

Returns the lagrange points for a CR3BP system.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

L: Langrange points requested, must be in range [1,5]

Outputs:

Tuple of Lagrange points

Throws ArgumentError if L is out of range [1,5]

References:

Rund, 2018

UnitfulAstrodynamics.ThreeBody.potential_energy_hessian — Function

Returns the partial derivative matrix of potential U.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

r: Non-dimensional position vector for the spacecraft.

Outputs:

Partial derivative matrix of potential U.

References:

Rund, 2018

UnitfulAstrodynamics.ThreeBody.accel — Function

accel(rₛ, vₛ, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:270.

Returns non-dimensional acceleration for CR3BP state.

UnitfulAstrodynamics.ThreeBody.RestrictedThreeBodySTMTic! — Function

RestrictedThreeBodySTMTic!(∂u, u, p, t)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:281.

Returns dynamics tic for combined Halo iterative solver state vector.

Arguments:

∂u: Derivative of state u.

u: State vector: [x, y, z, ẋ, ẏ, ż, Φ₁, Φ₂, Φ₃, Φ₄, Φ₅, Φ₆].

p: Parameters (contains non-dimensional mass parameter μ, positions x₁, x₂, and configuration).

t: Time in seconds.

Outputs:

None (sets derivative ∂u in-place).

References:

Rund, 2018.

UnitfulAstrodynamics.ThreeBody.state_transition_dynamics — Function

state_transition_dynamics(μ, r)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/Halo.jl:38.

Returns the derivative mapping of CR3BP state transition matrix, F.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

r: Non-dimensional position vector for the spacecraft.

Outputs:

Linear mapping from Φ to Φ̇, F.

References:

Rund, 2018

UnitfulAstrodynamics.ThreeBody.nondimensional_radius — Function

nondimensional_radius(r)
nondimensional_radius(r, xᵢ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/ThreeBody/ThreeBodyCalculations.jl:186.

Returns the spacecraft's nondimensional position w.r.t. body 1 (or 2).