ThreeBody Data Structures

The ThreeBody module contains one key structure: ThreeBodySystem. A ThreeBodySystem holds the nondimensional state of the spacecraft within the Circular Restricted Three-body Problem.

Missing docstring.

Missing docstring for RestrictedThreeBodySystem. Check Documenter's build log for details.

UnitfulAstrodynamics.ThreeBody.ThreeBodyState — Type

Describes a dimensional state of a spacecraft within the Circular Restriested Three-body Problem in the Synodic frame.

UnitfulAstrodynamics.ThreeBody.NondimensionalThreeBodyState — Type

Describes the non-dimensional state of a spacecraft within the Circular Restricted Three-body Problem in the Synodic frame.