TwoBody Data Structures

The TwoBody module contains two key structures: Orbit and CelestialBody.

An Orbit is the core structure for TwoBody calculations. It contains an orbital state (both Cartesian and the equivalent Keplerian representation), and a central body.

The central body within Orbit is of type CelestialBody. Common bodies in our solar system have been added for convenience, as described in Default CelestialBodies, but you can also make your own.

UnitfulAstrodynamics.TwoBody.TwoBodyState — Type

Struct for storing TwoBody Cartesian states for all conics.

UnitfulAstrodynamics.TwoBody.KeplerianState — Type

Struct for storing Keplerian states for all conics.

UnitfulAstrodynamics.TwoBody.Orbit — Function

Orbit(r, v, body)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/TwoBodyStates.jl:151.

Alias for TwoBodyState.

UnitfulAstrodynamics.TwoBody.CelestialBody — Type

Type representing large bodies in space. Currently, the following solar system bodies are supported:

Sun, Mercury, Venus, Earth, Moon (Luna), Mars, Jupiter, Saturn, Uranus, Neptune, Pluto.

Abstract Types and Pre-defined Parameters

The first section in TwoBody documentation described the core TwoBody Data Structures. Each data structure has an abstract parent type. All Orbit structures extend TwoBodySystem. In addition, all Orbit structures are paremeterized by their conic section, which is of type AbstractConic. All conic sections are pre-defined structures: Circular, Elliptical, Parabolic, Hyperbolic, and the Invalid conic is used to describe invalid orbital states (such as providing a NaN value to an Orbit constructor).

UnitfulAstrodynamics.TwoBody.RestrictedTwoBodySystem — Type

Abstract type for all two-body orbital representations.

UnitfulAstrodynamics.TwoBody.AbstractConic — Type

Abstract type for all four conic sections.

UnitfulAstrodynamics.TwoBody.Circular — Type

Type for orbits in the circular conic section.

UnitfulAstrodynamics.TwoBody.Elliptical — Type

Type for orbits in the elliptical conic section.

UnitfulAstrodynamics.TwoBody.Parabolic — Type

Type for orbits in the parabolic conic section.

UnitfulAstrodynamics.TwoBody.Hyperbolic — Type

Type for orbits in the hyperbolic conic section.

UnitfulAstrodynamics.TwoBody.Invalid — Type

Type for invalid orbits (orbits with NaN fields)

Default CelestialBodies

For convenence, the following common bodies in our solar system have already been defined!

Sun

Mercury

Venus

Earth

Moon

Luna

Mars

Jupiter

Saturn

Uranus

Neptune

Pluto