Circular Restricted Three Body Equations

All three-body equations!

AstrodynamicalCalculations.CR3BPCalculations — Module

Common calculations within Circular Restricted Three Body Problem dynamics.

Extended Help

Imports

Base

Core

DocStringExtensions

LinearAlgebra

Roots

StaticArrays

Exports

distance_scaling

distance_to_primary

distance_to_secondary

jacobi_constant

lagrange_point

nondimensional

nondimensional_radii

potential_energy

primary_synodic_position

redimensioned

reduced_mass

richardson_halo

secondary_synodic_position

time_scaling

velocity_scaling

zero_velocity_curves

AstrodynamicalCalculations.CR3BPCalculations.convergent_direction — Method

convergent_direction(monodromy; atol)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Calculates the eigenvector associated with the stable manifold of a Monodromy matrix.

AstrodynamicalCalculations.CR3BPCalculations.distance_scaling — Method

distance_scaling(a)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

The length scale factor used to nondimensionalize CR3BP states.

AstrodynamicalCalculations.CR3BPCalculations.distance_to_primary — Method

distance_to_primary(r, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns synodic distance to primary body.

AstrodynamicalCalculations.CR3BPCalculations.distance_to_secondary — Method

distance_to_secondary(r, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns synodic distance to secondary body.

AstrodynamicalCalculations.CR3BPCalculations.divergent_direction — Method

divergent_direction(monodromy; atol)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Calculates the direction associated with the unstable manifold of a Monodromy matrix.

AstrodynamicalCalculations.CR3BPCalculations.inertial_to_synodic — Method

inertial_to_synodic(rᵢ, t, ω)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Given a barycentric-inertial Cartesian state, returns the state in the synodic (rotating) reference frame.

AstrodynamicalCalculations.CR3BPCalculations.jacobi_constant — Method

jacobi_constant(r, v, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns the Jacobi Constant, C in the Synodic frame with Normalized units.

AstrodynamicalCalculations.CR3BPCalculations.lagrange_point — Function

lagrange_point(μ)
lagrange_point(μ, L)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns the lagrange points for a CR3BP system.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

L: Langrange points requested, must be in range [1,5]

Outputs:

Tuple of Lagrange points

Throws ArgumentError if L is out of range [1,5]

References:

Rund, 2018

AstrodynamicalCalculations.CR3BPCalculations.nondimensional — Method

nondimensional(r, v, t, a, μ₁, μ₂)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Normalizes a CR3BP orbit in the rotating reference frame.

AstrodynamicalCalculations.CR3BPCalculations.nondimensional_radii — Method

nondimensional_radii(r, xᵢ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns the spacecraft's nondimensional position w.r.t. body 1 (or 2).

AstrodynamicalCalculations.CR3BPCalculations.perturb_halo — Method

perturb_halo(r, v, perturbation; eps)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Perturbs a Cartesian state along a halo orbit onto the provided direction of the provided manifold.

AstrodynamicalCalculations.CR3BPCalculations.potential_energy — Method

potential_energy(r, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns the potential energy U in the Synodic frame with Normalized units.

AstrodynamicalCalculations.CR3BPCalculations.primary_synodic_position — Method

primary_synodic_position(μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Position of primary body.

AstrodynamicalCalculations.CR3BPCalculations.redimensioned — Method

redimensioned(rₙ, vₙ, tₙ, μ, a, Tₛ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Redimensionalizes a CR3BP orbit in the rotating reference frame.

AstrodynamicalCalculations.CR3BPCalculations.reduced_mass — Method

reduced_mass(μ₁, μ₂)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Return the reduced mass.

AstrodynamicalCalculations.CR3BPCalculations.richardson_halo — Method

richardson_halo(μ, L; Z, hemisphere, ϕ, steps)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns an analytical solution for a Halo orbit about L.

Extended Help

Arguments

μ: Non-dimensional mass parameter for the CR3BP system.

L: Lagrange point to orbit (L1 or L2).

Z: Desired non-dimensional Z-amplitude for Halo orbit.

hemisphere: Specifies northern or southern Halo orbit.

ϕ: Desired Halo orbit phase.

steps: Number of non-dimensional timepoints in returned state.

Outputs

Synodic position vectors r::Array{<:AbstractFloat}

Synodic velocity vectors v::Array{<:Abstractfloat}.

Halo orbit period Τ.

Throws ArgumentError if L is not 1 or 2.

References:

Rund, 2018.

AstrodynamicalCalculations.CR3BPCalculations.secondary_synodic_position — Method

secondary_synodic_position(μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Position of primary body.

AstrodynamicalCalculations.CR3BPCalculations.synodic_to_inertial — Method

synodic_to_inertial(rₛ, t, ω)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Given the Synodic frame vector, returns the vector in the barycentric-inertial reference frame.

AstrodynamicalCalculations.CR3BPCalculations.time_scaling — Method

time_scaling(a, μ₁, μ₂)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

The time scale factor used to nondimensionalize CR3BP states.

AstrodynamicalCalculations.CR3BPCalculations.velocity_scaling — Method

velocity_scaling(a, μ₁, μ₂)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

The velocity scale factor used to nondimensionalize CR3BP states.

AstrodynamicalCalculations.CR3BPCalculations.zero_velocity_curves — Method

zero_velocity_curves(r, v, μ; nondimensional_range)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Returns a Vector of Matrix values. Each Matrix contains a 3-column nondimensional position trajectory in the Synodic frame which represents a Zero Velocity Curve.