CommonTypes

The following abstract types are defined, which are common parent types for all submodules within Astrodynamics.

UnitfulAstrodynamics.CommonTypes.AbstractBody — Type

Abstract type for bodies in space: both CelestialBodys (in TwoBody.jl), and Bodys (in NBody.jl).

UnitfulAstrodynamics.CommonTypes.OrbitalSystem — Type

Abstract type describing all states in select Astrodynamics problems.

UnitfulAstrodynamics.CommonTypes.PropagationResult — Type

Abstract type describing a collection of states resulting from