ThreeBody Calculations

Common ThreeBody problem calculations are provided through functions.

Frame Representations

You can convert between the Inertial and Rotating (Synodic) reference frames through the inertial and synodic functions.

UnitfulAstrodynamics.Orbits.inertial — Function

Given the Synodic frame vector, returns the vector in the inertial reference frame.

Missing docstring.

Missing docstring for synodic. Check Documenter's build log for details.

Dimensionalization

Functions to nondimensionalize spacecraft states, and re-dimensionalize spacecraft states are provided.

UnitfulAstrodynamics.Orbits.time_scale_factor — Function

Returns time scale factor, Tₛ.

UnitfulAstrodynamics.Orbits.nondimensionalize_length — Function

Returns nondimensional length unit, DU.

UnitfulAstrodynamics.Orbits.nondimensionalize_velocity — Function

Returns nondimensional velocity unit, DU/DT.

UnitfulAstrodynamics.Orbits.nondimensionalize_time — Function

Returns nondimensional time unit, DT.

Missing docstring.

Missing docstring for nondimensionalize_mass_parameter. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.nondimensionalize — Function

Returns nondimensional form of (Unitful) scalar posiion.

Returns nondimensional form of (Unitful) position vector.

Returns nondimensional form of (Unitful) scalar velocity.

Returns nondimensional form of (Unitful) velocity vector.

Returns nondimensional form of (Unitful) velocity vector.

Returns nondimensional form of (Unitful) time duration.

Returns nondimensional form of (Unitful) time duration.

Returns nondimensional form of (Unitful) graviational parameters.

Returns nondimensional Circular Restricted Three-body State.

UnitfulAstrodynamics.Orbits.redimensionalize_length — Function

Returns dimensional length units.

UnitfulAstrodynamics.Orbits.redimensionalize_velocity — Function

Returns dimensional velocity units.

UnitfulAstrodynamics.Orbits.redimensionalize_time — Function

Returns dimensional time unit.

Returns dimensional time unit.

UnitfulAstrodynamics.Orbits.redimensionalize — Function

Returns dimensional (inertial) form of (Unitful) scalar posiion.

Returns dimensional (inertial) form of (Unitful) velocity vector.

Returns dimensional (inertial) form of (Unitful) time duration.

Returns dimensional (inertial) form of (Unitful) time duration.

Halo Orbit Solvers

Missing docstring.

Missing docstring for analyticalhalo. Check Documenter's build log for details.

Missing docstring.

Missing docstring for halo. Check Documenter's build log for details.

Missing docstring.

Missing docstring for monodromy. Check Documenter's build log for details.

Other Common Calculations

Missing docstring.

Missing docstring for potential_energy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for jacobi_constant. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.lagrange — Function

Returns the lagrange points for a CR3BP system.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

L: Langrange points requested, must be in range [1,5]

Outputs:

Tuple of Lagrange points

Throws ArgumentError if L is out of range [1,5]

References:

Rund, 2018

Missing docstring.

Missing docstring for potential_energy_hessian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for accel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RestrictedThreeBodySTMTic!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for state_transition_dynamics. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nondimensional_radius. Check Documenter's build log for details.