ThreeBody Calculations

Common ThreeBody problem calculations are provided through functions.

Frame Representations

You can convert between the Inertial and Rotating (Synodic) reference frames through the inertial and synodic functions.

UnitfulAstrodynamics.Orbits.inertial — Function

Given the Synodic frame vector, returns the vector in the inertial reference frame.

Given a SynodicCartesianState, returns the state in the inertial reference frame.

If given an inertial CartesianState, no operation needed.

Given a CR3BP orbit, returns the orbit in the inertial reference frame.

UnitfulAstrodynamics.Orbits.synodic — Function

Given an InertialCartesianState, returns the state in the synodic (rotating) reference frame.

If given a Synodic CartesianState, no operation needed.

Given a CR3BP orbit, returns the orbit in the Synodic (rotating) reference frame.

Dimensionalization

Functions to nondimensionalize spacecraft states, and re-dimensionalize spacecraft states are provided.

UnitfulAstrodynamics.Orbits.time_scale_factor — Function

Returns time scale factor, Tₛ.

UnitfulAstrodynamics.Orbits.nondimensionalize_length — Function

Returns nondimensional length unit, DU.

UnitfulAstrodynamics.Orbits.nondimensionalize_velocity — Function

Returns nondimensional velocity unit, DU/DT.

UnitfulAstrodynamics.Orbits.nondimensionalize_time — Function

Returns nondimensional time unit, DT.

Missing docstring.

Missing docstring for nondimensionalize_mass_parameter. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.nondimensionalize — Function

Returns nondimensional form of (Unitful) scalar posiion.

Returns nondimensional form of (Unitful) position vector.

Returns nondimensional form of (Unitful) scalar velocity.

Returns nondimensional form of (Unitful) velocity vector.

Returns nondimensional form of (Unitful) velocity vector.

Returns nondimensional form of (Unitful) time duration.

Returns nondimensional form of (Unitful) time duration.

Returns nondimensional form of (Unitful) graviational parameters.

Returns nondimensional Circular Restricted Three-body State.

UnitfulAstrodynamics.Orbits.redimensionalize_length — Function

Returns dimensional length units.

UnitfulAstrodynamics.Orbits.redimensionalize_velocity — Function

Returns dimensional velocity units.

UnitfulAstrodynamics.Orbits.redimensionalize_time — Function

Returns dimensional time unit.

Returns dimensional time unit.

UnitfulAstrodynamics.Orbits.redimensionalize — Function

Returns dimensional (inertial) form of (Unitful) scalar posiion.

Returns dimensional (inertial) form of (Unitful) velocity vector.

Returns dimensional (inertial) form of (Unitful) time duration.

Returns dimensional (inertial) form of (Unitful) time duration.

Un-normalizes (or re-dimensionalizes) a NormalizedCartesianState back into a CartesianState.

Un-normalizes (or re-dimensionalizes) a NormalizedCartesianState back into a CartesianState.

Halo Orbit Solvers

UnitfulAstrodynamics.Orbits.analyticalhalo — Function

Returns an analytical solution for a Halo orbit about L.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

Az: Desired non-dimensional Z-amplitude for Halo orbit.

ϕ: Desired Halo orbit phase.

steps: Number of non-dimensional timepoints in returned state.

L: Lagrange point to orbit (L1 or L2).

hemisphere: Specifies northern or southern Halo orbit.

Outputs:

Synodic position vector r::Array{<:AbstractFloat}

Synodic velocity vector v::Array{<:Abstractfloat}.

Halo orbit period Τ.

Throws ArgumentError if L is not 1 or 2.

References:

Rund, 2018.

UnitfulAstrodynamics.Propagators.halo — Function

Returns a numerical solution for a Halo orbit about L.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

Az: Desired non-dimensional Z-amplitude for Halo orbit.

ϕ: Desired Halo orbit phase.

L: Lagrange point to orbit (L1 or L2).

hemisphere: Specifies northern or southern Halo orbit.

Outputs:

Tuple of initial states: (r, v) where r::Vector{<:AbstractFloat}, v::Vector{<:Abstractfloat}.

Throws ArgumentError if L is not :L1 or :L2

References:

Rund, 2018.

A halo wrapper! Returns a CircularRestrictedThreeBodyOrbit. Returns a tuple: halo_orbit, halo_period.

UnitfulAstrodynamics.Propagators.monodromy — Function

Returns the Monodromy Matrix for a Halo orbit.

Other Common Calculations

UnitfulAstrodynamics.Orbits.potential_energy — Function

Returns the potential energy U in the Synodic frame with Normalized units.

Returns the potential energy U in the Synodic frame with Normalized units.

UnitfulAstrodynamics.Orbits.jacobi_constant — Function

Returns the Jacobi Constant, C in the Synodic frame with Normalized units.

Returns the Jacobi Constant, C in the Synodic frame with Normalized units.

UnitfulAstrodynamics.Orbits.lagrange — Function

Returns the lagrange points for a CR3BP system.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

L: Langrange points requested, must be in range [1,5]

Outputs:

Tuple of Lagrange points

Throws ArgumentError if L is out of range [1,5]

References:

Rund, 2018

UnitfulAstrodynamics.Propagators.potential_energy_hessian — Function

Returns the partial derivative matrix of potential U.

Arguments:

μ: Non-dimensional mass parameter for the CR3BP system.

r: Non-dimensional position vector for the spacecraft.

Outputs:

Partial derivative matrix of potential U.

References:

Rund, 2018

UnitfulAstrodynamics.Orbits.accel — Function

Returns non-dimensional acceleration for CR3BP state.

Missing docstring.

Missing docstring for RestrictedThreeBodySTMTic!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for state_transition_dynamics. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nondimensional_radius. Check Documenter's build log for details.