ThreeBody Calculations

Common ThreeBody problem calculations are provided through functions.

Frame Representations

You can convert between the Inertial and Rotating (Synodic) reference frames through the inertial and synodic functions.

UnitfulAstrodynamics.ThreeBody.inertial — Function

inertial(vecₛ, t)
inertial(vecₛ, t, ω)

Given the Synodic frame vector, returns the vector in the inertial reference frame.

UnitfulAstrodynamics.ThreeBody.synodic — Function

synodic(rᵢ, vᵢ, a, Tₛ)

Returns the position and velocity vectors in the synodic (rotating) reference frame.

Dimensionalization

Functions to nondimensionalize spacecraft states, and re-dimensionalize spacecraft states are provided.

UnitfulAstrodynamics.ThreeBody.time_scale_factor — Function

time_scale_factor(a, μ₁, μ₂)

Returns time scale factor, Tₛ.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_length — Function

nondimensionalize_length(rᵢ, a)

Returns nondimensional length unit, DU.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_velocity — Function

nondimensionalize_velocity(vᵢ, a, Tₛ)

Returns nondimensional velocity unit, DU/DT.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_time — Function

nondimensionalize_time(t, a, μ₁, μ₂)

Returns nondimensional time unit, DT.

UnitfulAstrodynamics.ThreeBody.nondimensionalize_mass_parameter — Function

nondimensionalize_mass_parameter(μ₁, μ₂)

Returns nondimensional mass parameter, μ.

UnitfulAstrodynamics.ThreeBody.nondimensionalize — Function

nondimensionalize(rᵢ, a)

Returns nondimensional form of (Unitful) scalar posiion.

nondimensionalize(rᵢ, a)

Returns nondimensional form of (Unitful) position vector.

nondimensionalize(vᵢ, a)

Returns nondimensional form of (Unitful) scalar velocity.

nondimensionalize(vᵢ, a, Tₛ)

Returns nondimensional form of (Unitful) velocity vector.

nondimensionalize(vᵢ, a, μ₁, μ₂)

Returns nondimensional form of (Unitful) velocity vector.

nondimensionalize(t, Tₛ)

Returns nondimensional form of (Unitful) time duration.

nondimensionalize(t, a, μ₁, μ₂)

Returns nondimensional form of (Unitful) time duration.

nondimensionalize(μ₁, μ₂)

Returns nondimensional form of (Unitful) graviational parameters.

nondimensionalize(r₃, v₃, Δt, μ₁, μ₂, a)

Returns nondimensional Circular Restricted Three-body State.

UnitfulAstrodynamics.ThreeBody.redimensionalize_length — Function

redimensionalize_length(rᵢ, a)

Returns dimensional length units.

UnitfulAstrodynamics.ThreeBody.redimensionalize_velocity — Function

redimensionalize_velocity(vᵢ, a, Tₛ)

Returns dimensional velocity units.

UnitfulAstrodynamics.ThreeBody.redimensionalize_time — Function

redimensionalize_time(t, a, μ₁, μ₂)

Returns dimensional time unit.

UnitfulAstrodynamics.ThreeBody.redimensionalize — Function

redimensionalize(rᵢ, a)

Returns dimensional (inertial) form of (Unitful) scalar posiion.

redimensionalize(rᵢ, a)

Returns dimensional (inertial) form of (Unitful) position vector.

redimensionalize(vᵢ, a)

Returns dimensional (inertial) form of (Unitful) scalar velocity.

redimensionalize(vᵢ, a, Tₛ)

Returns dimensional (inertial) form of (Unitful) velocity vector.

redimensionalize(vᵢ, a, μ₁, μ₂)

Returns dimensional (inertial) form of (Unitful) velocity vector.

redimensionalize(t, Tₛ)

Returns dimensional (inertial) form of (Unitful) time duration.

redimensionalize(t, a, μ₁, μ₂)

Returns dimensional (inertial) form of (Unitful) time duration.

Other Common Calculations

UnitfulAstrodynamics.ThreeBody.potential_energy — Function

potential_energy(r, μ, x₁, x₂)

Returns the potential energy U.

UnitfulAstrodynamics.ThreeBody.jacobi_constant — Function

jacobi_constant(r, v, μ, x₁, x₂)

Returns the Jacobi Constant C.

UnitfulAstrodynamics.ThreeBody.nondimensional_radius — Function

nondimensional_radius(r)
nondimensional_radius(r, xᵢ)

Returns the spacecraft's nondimensional position w.r.t. body 1 (or 2).