TwoBody Calculations

For convenience, common TwoBody problem calculations are provided through functions. Often, these functions are provided with common arguments (such as orbital_period(a,μ)), and with TwoBody Data Structures arguments (such as orbital_period(::Orbit)).

Orbital Representations

You can convert between Cartesian and Keplerian TwoBody orbital representations by using cartesian and keplerian.

UnitfulAstrodynamics.TwoBody.cartesian — Function

cartesian(e, a, i, Ω, ω, ν, μ)
cartesian(e, a, i, Ω, ω, ν, body)

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

cartesian(i, Ω, ω, rₚ, vₚ)

Returns a Cartesian (inertial) representation of the provied Perifocal state.

UnitfulAstrodynamics.TwoBody.keplerian — Function

keplerian(rᵢ, vᵢ, μ)
keplerian(rᵢ, vᵢ, body)

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

UnitfulAstrodynamics.TwoBody.perifocal — Function

perifocal(a, e, ν, μ)

Returns position and velocity vectors in the Perifocal frame.

Kepler's Prediction Problem

The function kepler can solve Kepler's Prediction Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.kepler — Function

kepler(orbit)
kepler(orbit, Δtᵢ; tol, max_iter)

Solves Kepler's Problem for orbit and Δtᵢ.

Lambert's Problem

The function lambert can solve Lambert's Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.lambert — Function

lambert(r̅₁, r̅₂, μ, Δt)
lambert(r̅₁, r̅₂, μ, Δt, trajectory; tol, max_iter)

Solves Lambert's problem through the use of univeral variables.

Common TwoBody Problem Calculations

UnitfulAstrodynamics.TwoBody.semimajor_axis — Function

semimajor_axis(r, v, μ)

Returns semimajor axis parameter, a.

UnitfulAstrodynamics.TwoBody.eccentricity — Function

eccentricity(rᵢ, vᵢ, μ)

Returns orbital eccentricity, e.

UnitfulAstrodynamics.TwoBody.eccentricity_vector — Function

eccentricity_vector(rᵢ, vᵢ, μ)

Returns orbital eccentricity vector e̅.

UnitfulAstrodynamics.TwoBody.inclination — Function

inclination(orbit)

Returns orbital inclination, i.

UnitfulAstrodynamics.TwoBody.true_anomoly — Function

true_anomoly(r, h, e, μ)

Returns true anomoly, ν.

UnitfulAstrodynamics.TwoBody.periapsis_radius — Function

periapsis_radius(a, e)

Returns periapsis radius, rₚ.

UnitfulAstrodynamics.TwoBody.apoapsis_radius — Function

apoapsis_radius(a, e)

Returns periapsis radius, rₐ.

UnitfulAstrodynamics.TwoBody.periapsis_velocity — Function

periapsis_velocity(orbit)

Returns periapsis velocity, vₚ, for any orbital representation.

UnitfulAstrodynamics.TwoBody.apoapsis_velocity — Function

apoapsis_velocity(orbit)

Returns apoapsis velocity, v_a, for any orbital representation.

UnitfulAstrodynamics.TwoBody.radius — Function

radius(p, e, ν)

Returns radius, r.

UnitfulAstrodynamics.TwoBody.velocity — Function

velocity(r, a, μ)

Returns instantaneous velocity, v, for any orbital representation.

UnitfulAstrodynamics.TwoBody.mass — Function

mass(body)

Returns mass m.

UnitfulAstrodynamics.TwoBody.mass_parameter — Function

mass_parameter(body)

Returns mass parameter μ.

UnitfulAstrodynamics.TwoBody.orbital_period — Function

orbital_period(a, μ)

Returns the orbital period.

UnitfulAstrodynamics.TwoBody.time_since_periapsis — Function

time_since_periapsis(n, e, E)

Returns time since periapsis, t.

UnitfulAstrodynamics.TwoBody.mean_motion — Function

mean_motion(a, μ)

Returns mean motion, n.

UnitfulAstrodynamics.TwoBody.mean_motion_vector — Function

mean_motion_vector(orbit)

Returns mean motion vector, n̄.

UnitfulAstrodynamics.TwoBody.semi_parameter — Function

semi_parameter(a, e)

Returns semilatus parameter, p.

Missing docstring.

Missing docstring for conic_anomoly. Check Documenter's build log for details.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum — Function

specific_angular_momentum(rᵢ, vᵢ)

Returns scalar specific angular momentum vector, h.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum_vector — Function

specific_angular_momentum_vector(rᵢ, vᵢ)

Returns specific angular momentum vector, h̅.

UnitfulAstrodynamics.TwoBody.specific_energy — Function

specific_energy(a, μ)

Returns specific orbital energy, ϵ.

UnitfulAstrodynamics.TwoBody.conic — Function

conic(e)

Returns the conic section, as specified by eccentricity e.

UnitfulAstrodynamics.TwoBody.isinvalid — Function

isinvalid(orbit)

Checks for NaN valued orbital states, which are used to indicate an invalid Orbit.