TwoBody Calculations

For convenience, common TwoBody problem calculations are provided through functions. Often, these functions are provided with common arguments (such as orbital_period(a,μ)), and with TwoBody Data Structures arguments (such as orbital_period(::Orbit)).

Orbital Representations

You can convert between Cartesian and Keplerian TwoBody orbital representations by using cartesian and keplerian.

UnitfulAstrodynamics.TwoBody.cartesian — Function

cartesian(e, a, i, Ω, ω, ν, μ)

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

UnitfulAstrodynamics.TwoBody.keplerian — Function

keplerian(rᵢ, vᵢ, body::CelestialBody)

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

UnitfulAstrodynamics.TwoBody.inertial — Function

inertial(i, Ω, ω, vec₃)
inertial(orbit::Orbit)

Transforms 3-vector from Perifocal frame to Cartesian space (x,y,z).

UnitfulAstrodynamics.TwoBody.perifocal — Function

perifocal(i, Ω, ω, vec₃)
perifocal(orbit::Orbit)

Transforms 3-vector from Cartesian frame to Perifocal frame.

Kepler's Prediction Problem

The function kepler can solve Kepler's Prediction Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.kepler — Function

kepler(orbit::T, Δtᵢ::N = orbital_period(orbit)) where {T<:Orbit, N<:Number}

Solves Kepler's Problem for orbit and Δtᵢ.

Common TwoBody Problem Calculations

UnitfulAstrodynamics.TwoBody.semimajor_axis — Function

semimajor_axis(r, v, μ)
semimajor_axis(orbit::Orbit)

Returns semimajor axis parameter, a.

UnitfulAstrodynamics.TwoBody.eccentricity — Function

eccentricity(rᵢ, vᵢ, μ)
eccentricity(orbit::Orbit)

Returns orbital eccentricity, e.

UnitfulAstrodynamics.TwoBody.eccentricity_vector — Function

eccentricity_vector(rᵢ, vᵢ, μ)
eccentricity_vector(orbit::Orbit)

Returns orbital eccentricity vector e̅.

UnitfulAstrodynamics.TwoBody.inclination — Function

inclination(orbit::Orbit)

Returns orbital inclination, i.

UnitfulAstrodynamics.TwoBody.true_anomoly — Function

true_anomoly(r, h, e, μ)

Returns true anomoly, ν.

UnitfulAstrodynamics.TwoBody.periapsis_radius — Function

periapsis_radius(a, e)
periapsis_radius(orbit::Orbit)

Returns periapsis radius, rᵢ_p.

UnitfulAstrodynamics.TwoBody.apoapsis_radius — Function

apoapsis_radius(a, e)
apoapsis_radius(orbit::Orbit)

Returns periapsis radius, r_a.

UnitfulAstrodynamics.TwoBody.periapsis_velocity — Function

periapsis_velocity(orbit::T) where T<:Orbit

Returns periapsis velocity, v_p, for any orbital representation.

UnitfulAstrodynamics.TwoBody.apoapsis_velocity — Function

apoapsis_velocity(orbit::T) where T<:Orbit

Returns apoapsis velocity, v_a, for any orbital representation.

UnitfulAstrodynamics.TwoBody.radius — Function

radius(p, e, ν)
radius(orbit::Orbit)

Returns instantaneous radius, r.

UnitfulAstrodynamics.TwoBody.velocity — Function

velocity(r, a, μ)
velocity(orbit::Orbit)

Returns instantaneous velocity, v, for any orbital representation.

UnitfulAstrodynamics.TwoBody.orbital_period — Function

orbital_period(a, μ)
orbital_period(orbit::Orbit)

Returns the orbital period.

UnitfulAstrodynamics.TwoBody.time_since_periapsis — Function

time_since_periapsis(n, e, E)
time_since_periapsis(orbit::Orbit)

Returns time since periapsis, t.

UnitfulAstrodynamics.TwoBody.mean_motion — Function

mean_motion(a, μ)
mean_motion(orbit::Orbit)

Returns mean motion, n.

UnitfulAstrodynamics.TwoBody.semi_parameter — Function

semi_parameter(a, e)
semi_parameter(orbit::Orbit)

Returns semilatus parameter, p.

UnitfulAstrodynamics.TwoBody.conic_anomoly — Function

conic_anomoly(orbit::Orbit{Elliptical})
conic_anomoly(orbit::Orbit{Hyperbolic})

Returns eccentric anomoly, E, parabolic anomoly, B, or hyperbolic anomoly, H.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum — Function

specific_angular_momentum(rᵢ, vᵢ)
specific_angular_momentum(orbit::Orbit)

Returns scalar specific angular momentum vector, h.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum_vector — Function

specific_angular_momentum_vector(rᵢ, vᵢ)
specific_angular_momentum_vector(orbit::Orbit)

Returns specific angular momentum vector, h̅.

UnitfulAstrodynamics.TwoBody.specific_energy — Function

specific_energy(a, μ)
specific_energy(r, v, μ)
specific_energy(orbit::Orbit)

Returns specific orbital energy, ϵ.

UnitfulAstrodynamics.TwoBody.conic — Function

conic(e::T) where T<:Number
conic(orbit::Orbit)

Returns the conic section, as specified by eccentricity e.

UnitfulAstrodynamics.TwoBody.isinvalid — Function

isinvalid(orbit::Orbit)

Checks for NaN valued orbital states, which are used to indicate an invalid Orbit.

Base.isapprox — Function

isapprox(::Orbit, ::Orbit; atol=1e-8)

Returns true if all elements in each system are within atol of the other.

Base.isequal — Function

isequal(::Orbit, ::Orbit)

Returns true if all elements of each system are identically equal.