TwoBody Calculations

For convenience, common TwoBody problem calculations are provided through functions. Often, these functions are provided with common arguments (such as orbital_period(a,μ)), and with TwoBody Data Structures arguments (such as orbital_period(::Orbit)).

Orbital Representations

You can convert between Cartesian and Keplerian TwoBody orbital representations by using cartesian and keplerian.

UnitfulAstrodynamics.TwoBody.cartesian — Function

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

Returns a Cartesian (inertial) representation of the provied Perifocal state.

UnitfulAstrodynamics.TwoBody.keplerian — Function

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

UnitfulAstrodynamics.TwoBody.perifocal — Function

Returns position and velocity vectors in the Perifocal frame.

Kepler's Prediction Problem

The function kepler can solve Kepler's Prediction Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.kepler — Function

Solves Kepler's Problem for orbit and Δtᵢ.

Implements Kepler's Prediction Problem for generic r, v.

Arguments:

r: Initial spacecraft position. Any abstract vector with length 3 and type <: Unitful.Length

v: Initial spacecraft velocity. Any abstract vector with length 3 and type <: Unitful.Velocity

μ: Central body mass parameter. Any abstract vector with length 3 and type <: Unitful.AbstractQuantity with units compatabile with km^3/s^2

Δtᵢ: Propagation time. Any scalar with type <: Unitful.Time

Lambert's Problem

The function lambert can solve Lambert's Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.lambert — Function

Solves Lambert's problem through the use of univeral variables.

Common TwoBody Problem Calculations

UnitfulAstrodynamics.TwoBody.semimajor_axis — Function

Returns semimajor axis parameter, a.

UnitfulAstrodynamics.TwoBody.eccentricity — Function

Returns orbital eccentricity, e.

UnitfulAstrodynamics.TwoBody.eccentricity_vector — Function

Returns orbital eccentricity vector e̅.

UnitfulAstrodynamics.TwoBody.inclination — Function

Returns orbital inclination, i.

UnitfulAstrodynamics.TwoBody.true_anomoly — Function

Returns true anomoly, ν.

Missing docstring.

Missing docstring for periapsis_radius. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apoapsis_radius. Check Documenter's build log for details.

Missing docstring.

Missing docstring for periapsis_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apoapsis_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for radius. Check Documenter's build log for details.

Missing docstring.

Missing docstring for radius_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for velocity. Check Documenter's build log for details.

UnitfulAstrodynamics.AstrodynamicsCore.velocity_vector — Function

Returns the Unitful velocity vector of the CartesianState.

Returns velocity vector, vᵢ.

UnitfulAstrodynamics.TwoBody.mass — Function

Returns mass m.

UnitfulAstrodynamics.TwoBody.mass_parameter — Function

Returns mass parameter μ.

UnitfulAstrodynamics.TwoBody.period — Function

Returns the orbital period.

UnitfulAstrodynamics.TwoBody.time_since_periapsis — Function

Returns time since periapsis, t.

UnitfulAstrodynamics.TwoBody.mean_motion — Function

Returns mean motion, n.

UnitfulAstrodynamics.TwoBody.mean_motion_vector — Function

Returns mean motion vector, n̄.

UnitfulAstrodynamics.TwoBody.semi_parameter — Function

Returns semilatus parameter, p.

UnitfulAstrodynamics.TwoBody.eccentric_anomoly — Function

Returns eccentric anomoly, E, parabolic anomoly, B, or hyperbolic anomoly, H.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum — Function

Returns scalar specific angular momentum vector, h.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum_vector — Function

Returns specific angular momentum vector, h̅.

UnitfulAstrodynamics.TwoBody.specific_energy — Function

Returns specific orbital energy, ϵ.

UnitfulAstrodynamics.TwoBody.conic — Function

Returns the conic section, as specified by eccentricity e.