TwoBody Calculations

For convenience, common TwoBody problem calculations are provided through functions. Often, these functions are provided with common arguments (such as orbital_period(a,μ)), and with TwoBody Data Structures arguments (such as orbital_period(::Orbit)).

Orbital Representations

You can convert between Cartesian and Keplerian TwoBody orbital representations by using cartesian and keplerian.

UnitfulAstrodynamics.Orbits.cartesian — Function

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

Returns a Cartesian (inertial) representation of the provied Perifocal state.

UnitfulAstrodynamics.Orbits.keplerian — Function

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

UnitfulAstrodynamics.Orbits.perifocal — Function

Returns position and velocity vectors in the Perifocal frame.

Kepler's Prediction Problem

The function kepler can solve Kepler's Prediction Problem for an Orbit.

UnitfulAstrodynamics.Orbits.kepler — Function

Solves Kepler's Problem for orbit and Δtᵢ.

Implements Kepler's Prediction Problem for generic r, v.

Arguments:

r: Initial spacecraft position. Any abstract vector with length 3 and type <: Unitful.Length

v: Initial spacecraft velocity. Any abstract vector with length 3 and type <: Unitful.Velocity

μ: Central body mass parameter. Any abstract vector with length 3 and type <: Unitful.AbstractQuantity with units compatabile with km^3/s^2

Δtᵢ: Propagation time. Any scalar with type <: Unitful.Time

Lambert's Problem

The function lambert can solve Lambert's Problem for an Orbit.

UnitfulAstrodynamics.Orbits.lambert — Function

Solves Lambert's problem through the use of univeral variables.

Common TwoBody Problem Calculations

UnitfulAstrodynamics.Orbits.semimajor_axis — Function

Returns Unitful Restricted Two-body orbit semimajor axis.

Returns semimajor axis parameter, a.

UnitfulAstrodynamics.Orbits.eccentricity — Function

Returns Unitful Restricted Two-body orbit eccentricity.

Returns orbital eccentricity, e.

UnitfulAstrodynamics.Orbits.eccentricity_vector — Function

Returns orbital eccentricity vector e̅.

UnitfulAstrodynamics.Orbits.inclination — Function

Returns Unitful Restricted Two-body orbit inclination.

Returns orbital inclination, i.

Returns orbital inclination, i.

UnitfulAstrodynamics.Orbits.true_anomoly — Function

Returns Unitful Restricted Two-body orbit true anomoly.

Returns true anomoly, ν.

Returns true anomoly, ν.

Returns true anomoly, ν.

UnitfulAstrodynamics.Orbits.periapsis_radius — Function

Returns periapsis scalar_position, rₚ.

Missing docstring.

Missing docstring for apoapsis_radius. Check Documenter's build log for details.

Missing docstring.

Missing docstring for periapsis_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for apoapsis_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for radius. Check Documenter's build log for details.

Missing docstring.

Missing docstring for radius_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for velocity. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.velocity_vector — Function

Returns the Unitful velocity vector of the CartesianState.

Returns velocity vector, vᵢ.

Returns the Unitful velocity vector of the CartesianState.

Missing docstring.

Missing docstring for mass. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.mass_parameter — Function

Returns mass parameter μ.

UnitfulAstrodynamics.Orbits.period — Function

Returns the orbital period.

Returns the orbital period.

Returns the orbital period.

Returns the orbital period.

Returns the orbital period.

UnitfulAstrodynamics.Orbits.time_since_periapsis — Function

Returns time since periapsis, t.

Returns time since periapsis, t.

UnitfulAstrodynamics.Orbits.mean_motion — Function

Returns mean motion, n.

UnitfulAstrodynamics.Orbits.mean_motion_vector — Function

Returns mean motion vector, n̄.

UnitfulAstrodynamics.Orbits.semi_parameter — Function

Returns semilatus parameter, p.

UnitfulAstrodynamics.Orbits.eccentric_anomoly — Function

Returns eccentric anomoly, E, parabolic anomoly, B, or hyperbolic anomoly, H.

UnitfulAstrodynamics.Orbits.specific_angular_momentum — Function

Returns scalar specific angular momentum vector, h.

UnitfulAstrodynamics.Orbits.specific_angular_momentum_vector — Function

Returns specific angular momentum vector, h̅.

UnitfulAstrodynamics.Orbits.specific_energy — Function

Returns specific orbital energy, ϵ.

UnitfulAstrodynamics.Orbits.conic — Function

Returns the conic section, as specified by eccentricity e.