TwoBody Calculations

For convenience, common TwoBody problem calculations are provided through functions. Often, these functions are provided with common arguments (such as orbital_period(a,μ)), and with TwoBody Data Structures arguments (such as orbital_period(::Orbit)).

Orbital Representations

You can convert between Cartesian and Keplerian TwoBody orbital representations by using cartesian and keplerian.

UnitfulAstrodynamics.TwoBody.cartesian — Function

cartesian(e, a, i, Ω, ω, ν, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:112.

cartesian(e, a, i, Ω, ω, ν, body)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:117.

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

UnitfulAstrodynamics.TwoBody.keplerian — Function

keplerian(rᵢ, vᵢ, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:73.

keplerian(rᵢ, vᵢ, body)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:101.

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

Missing docstring.

Missing docstring for inertial. Check Documenter's build log for details.

UnitfulAstrodynamics.TwoBody.perifocal — Function

perifocal(a, e, ν, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:151.

Returns position and velocity vectors in the Perifocal frame.

Kepler's Prediction Problem

The function kepler can solve Kepler's Prediction Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.kepler — Function

kepler(orbit)
kepler(orbit, Δtᵢ; tol, max_iter)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/kepler.jl:12.

Solves Kepler's Problem for orbit and Δtᵢ.

Lambert's Problem

The function lambert can solve Lambert's Problem for an Orbit.

UnitfulAstrodynamics.TwoBody.lambert — Function

lambert(r̅₁, r̅₂, μ, Δt)
lambert(r̅₁, r̅₂, μ, Δt, trajectory; tol, max_iter)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/lambert.jl:14.

Solves Lambert's problem through the use of univeral variables.

Common TwoBody Problem Calculations

UnitfulAstrodynamics.TwoBody.semimajor_axis — Function

semimajor_axis(r, v, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:171.

Returns semimajor axis parameter, a.

UnitfulAstrodynamics.TwoBody.eccentricity — Function

eccentricity(rᵢ, vᵢ, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:214.

Returns orbital eccentricity, e.

UnitfulAstrodynamics.TwoBody.eccentricity_vector — Function

eccentricity_vector(rᵢ, vᵢ, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:205.

Returns orbital eccentricity vector e̅.

UnitfulAstrodynamics.TwoBody.inclination — Function

inclination(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:344.

Returns orbital inclination, i.

UnitfulAstrodynamics.TwoBody.true_anomoly — Function

true_anomoly(r, h, e, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:288.

Returns true anomoly, ν.

UnitfulAstrodynamics.TwoBody.periapsis_radius — Function

periapsis_radius(a, e)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:239.

Returns periapsis radius, rₚ.

UnitfulAstrodynamics.TwoBody.apoapsis_radius — Function

apoapsis_radius(a, e)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:245.

Returns periapsis radius, rₐ.

UnitfulAstrodynamics.TwoBody.periapsis_velocity — Function

periapsis_velocity(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:253.

Returns periapsis velocity, vₚ, for any orbital representation.

UnitfulAstrodynamics.TwoBody.apoapsis_velocity — Function

apoapsis_velocity(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:262.

Returns apoapsis velocity, v_a, for any orbital representation.

UnitfulAstrodynamics.TwoBody.radius — Function

radius(p, e, ν)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:226.

Returns radius, r.

UnitfulAstrodynamics.TwoBody.velocity — Function

velocity(r, a, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:233.

Returns instantaneous velocity, v, for any orbital representation.

UnitfulAstrodynamics.TwoBody.mass — Function

mass(body)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:272.

Returns mass m.

UnitfulAstrodynamics.TwoBody.mass_parameter — Function

mass_parameter(body)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:277.

Returns mass parameter μ.

UnitfulAstrodynamics.TwoBody.orbital_period — Function

orbital_period(a, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:282.

Returns the orbital period.

UnitfulAstrodynamics.TwoBody.time_since_periapsis — Function

time_since_periapsis(n, e, E)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:334.

Returns time since periapsis, t.

UnitfulAstrodynamics.TwoBody.mean_motion — Function

mean_motion(a, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:294.

Returns mean motion, n.

UnitfulAstrodynamics.TwoBody.mean_motion_vector — Function

mean_motion_vector(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:304.

Returns mean motion vector, n̄.

UnitfulAstrodynamics.TwoBody.semi_parameter — Function

semi_parameter(a, e)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:220.

Returns semilatus parameter, p.

UnitfulAstrodynamics.TwoBody.conic_anomoly — Function

conic_anomoly(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:316.

Returns eccentric anomoly, E, parabolic anomoly, B, or hyperbolic anomoly, H.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum — Function

specific_angular_momentum(rᵢ, vᵢ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:190.

Returns scalar specific angular momentum vector, h.

UnitfulAstrodynamics.TwoBody.specific_angular_momentum_vector — Function

specific_angular_momentum_vector(rᵢ, vᵢ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:181.

Returns specific angular momentum vector, h̅.

UnitfulAstrodynamics.TwoBody.specific_energy — Function

specific_energy(a, μ)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:196.

Returns specific orbital energy, ϵ.

UnitfulAstrodynamics.TwoBody.conic — Function

conic(e)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:12.

Returns the conic section, as specified by eccentricity e.

UnitfulAstrodynamics.TwoBody.isinvalid — Function

isinvalid(orbit)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_states.jl:205.

Checks for NaN valued orbital states, which are used to indicate an invalid Orbit.

Base.isapprox — Function

isapprox(c1, c2; atol)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:351.

Returns true if all elements in each system are within atol of the other.

Base.isequal — Function

isequal(c1, c2)

defined at /home/runner/work/UnitfulAstrodynamics.jl/UnitfulAstrodynamics.jl/src/TwoBody/twobody_calculations.jl:368.

Returns true if all elements of each system are identically equal.