TwoBody Data Structures

The TwoBody module contains two key structures: Orbit and CelestialBody.

An Orbit is the core structure for TwoBody calculations. It contains an orbital state (both Cartesian and the equivalent Keplerian representation), and a central body.

The central body within Orbit is of type CelestialBody. Common bodies in our solar system have been added for convenience, as described in Default CelestialBodies, but you can also make your own.

UnitfulAstrodynamics.TwoBody.Orbit — Type

Orbit{T<:AbstractConic}

Struct for storing TwoBody orbital states for all conics.

UnitfulAstrodynamics.TwoBody.CelestialBody — Type

struct CelestialBody(m, R, μ)
CelestialBody(m, R) = CelestialBody(m, R, G * m)

Type representing large bodies in space. Currently, the following solar system bodies are supported:

Sun, Mercury, Venus, Earth, Moon (Luna), Mars, Jupiter, Saturn, Uranus, Neptune, Pluto.

Abstract Types and Pre-defined Parameters

The first section in TwoBody documentation described the core TwoBody Data Structures. Each data structure has an abstract parent type. All Orbit structures extend TwoBodySystem. In addition, all Orbit structures are paremeterized by their conic section, which is of type AbstractConic. All conic sections are pre-defined structures: Circular, Elliptical, Parabolic, Hyperbolic, and the Invalid conic is used to describe invalid orbital states (such as providing a NaN value to an Orbit constructor).

UnitfulAstrodynamics.TwoBody.TwoBodySystem — Type

TwoBodyState

Abstract type for all two-body orbital representations.

UnitfulAstrodynamics.TwoBody.AbstractConic — Type

AbstractConic

Abstract type for all four conic sections.

UnitfulAstrodynamics.TwoBody.Circular — Type

Circular

Type for orbits in the circular conic section.

UnitfulAstrodynamics.TwoBody.Elliptical — Type

Elliptical

Type for orbits in the elliptical conic section.

UnitfulAstrodynamics.TwoBody.Parabolic — Type

Parabolic

Type for orbits in the parabolic conic section.

UnitfulAstrodynamics.TwoBody.Hyperbolic — Type

Hyperbolic

Type for orbits in the hyperbolic conic section.

UnitfulAstrodynamics.TwoBody.Invalid — Type

Invalid

Type for invalid orbits (orbits with NaN fields)

UnitfulAstrodynamics.TwoBody.InvalidOrbit — Function

InvalidOrbit(body::CelestialBody)

Returns a Orbit with NaN state values. Used by propagate_twobody and kepler to indicate failed convergance.

Default CelestialBodies

For convenence, the following common bodies in our solar system have already been defined!

Sun

Mercury

Venus

Earth

Moon

Luna

Mars

Jupiter

Saturn

Uranus

Neptune

Pluto