TwoBody Data Structures

The TwoBody module contains two key structures: Orbit and CelestialBody.

An Orbit is the core structure for TwoBody calculations. It contains an orbital state (both Cartesian and the equivalent Keplerian representation), and a central body.

The central body within Orbit is of type CelestialBody. Common bodies in our solar system have been added for convenience, as described in Default CelestialBodies, but you can also make your own.

Missing docstring.

Missing docstring for TwoBodyState. Check Documenter's build log for details.

UnitfulAstrodynamics.Orbits.KeplerianState — Type

Struct for storing Keplerian states.

UnitfulAstrodynamics.Orbits.Orbit — Function

RestrictedTwoBodyOrbit instanes are likely the most commonly used feature of this package. To accomadate this, Orbit is an alias for common RestrictedTwoBodyOrbit constructers.

RestrictedTwoBodyOrbit instanes are likely the most commonly used feature of this package. To accomadate this, Orbit is an alias for common RestrictedTwoBodyOrbit constructers.

RestrictedTwoBodyOrbit instanes are likely the most commonly used feature of this package. To accomadate this, Orbit is an alias for common RestrictedTwoBodyOrbit constructers.

RestrictedTwoBodyOrbit instanes are likely the most commonly used feature of this package. To accomadate this, Orbit is an alias for common RestrictedTwoBodyOrbit constructers.

A simpler constructor for a Synodic CR3BP state with or without the local linearization (state transition matrix) included.

Please note that this is a convenience constructor for quick calculations. This method is not type stable, since the kwarg include_stm directly affects the output type of the function. This sub-optimal code is great for quickly constructing an initial state, but this method will be slower than an explicit CircularRestrictedThreeBodyOrbit constructor.

Missing docstring.

Missing docstring for CelestialBody. Check Documenter's build log for details.

Abstract Types and Pre-defined Parameters

The first section in TwoBody documentation described the core TwoBody Data Structures. Each data structure has an abstract parent type. All Orbit structures extend TwoBodySystem. In addition, all Orbit structures are paremeterized by their conic section, which is of type AbstractConic. All conic sections are pre-defined structures: Circular, Elliptical, Parabolic, Hyperbolic, and the Invalid conic is used to describe invalid orbital states (such as providing a NaN value to an Orbit constructor).

UnitfulAstrodynamics.Orbits.RestrictedTwoBodySystem — Type

Type representing large bodies in space. Currently, the following solar system bodies are supported:

Sun, Mercury, Venus, Earth, Moon (Luna), Mars, Jupiter, Saturn, Uranus, Neptune, Pluto.

UnitfulAstrodynamics.Orbits.AbstractConic — Type

Abstract type for all four conic sections.

UnitfulAstrodynamics.Orbits.Circular — Type

Type for orbits in the circular conic section.

UnitfulAstrodynamics.Orbits.Elliptical — Type

Type for orbits in the elliptical conic section.

UnitfulAstrodynamics.Orbits.Parabolic — Type

Type for orbits in the parabolic conic section.

UnitfulAstrodynamics.Orbits.Hyperbolic — Type

Type for orbits in the hyperbolic conic section.

UnitfulAstrodynamics.Orbits.Invalid — Type

Type for invalid orbits (orbits with NaN fields)

Default CelestialBodies

For convenence, the following common bodies in our solar system have already been defined!

Sun

Mercury

Venus

Earth

Moon

Luna

Mars

Jupiter

Saturn

Uranus

Neptune

Pluto