Circular Restricted Three Body Solvers

All three-body solvers!

AstrodynamicalSolvers.CR3BSolvers — Module

Solvers specific to the Circular Restricted Three Body Problem.

Extended Help

Exports

halo

lyapunov

Imports

AstrodynamicalCalculations

AstrodynamicalModels

Base

Core

DocStringExtensions

LinearAlgebra

ModelingToolkit

OrdinaryDiffEqVerner

StaticArrays

AstrodynamicalSolvers.CR3BSolvers.extraplanar_differential — Method

extraplanar_differential(state, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Given a full state vector for CR3BP dynamics, including vertically concatenated columns of the state transition matrix, return the differential correction term for a periodic orbit.

AstrodynamicalSolvers.CR3BSolvers.halo — Method

halo(x, z, ẏ, μ, T; reltol, abstol, maxiters)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Iterate on an initial guess for halo orbit conditions.

AstrodynamicalSolvers.CR3BSolvers.halo — Method

halo(μ, lagrange; amplitude, phase, hemisphere, kwargs...)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Given a nondimensional mass parameter μ, and orbit characteristics, construct an initial guess using Richardson's analytical solution, and iterate on that guess using a differential corrector.

AstrodynamicalSolvers.CR3BSolvers.lyapunov — Method

lyapunov(x, ẏ, μ, T; reltol, abstol, maxiters)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Iterate on an initial guess for Lyapunov orbit conditions.

AstrodynamicalSolvers.CR3BSolvers.planar_differential — Method

planar_differential(state, μ)

CR3BP Dynamics

This computation is valid for Circular Restricted Three Body Problem dynamics.

Given a full state vector for CR3BP dynamics, including vertically concatenated columns of the state transition matrix, return the differential correction term for a planar periodic orbit.