Docstrings

Documentation for all types, and functions in GeneralAstrodynamics.

GeneralAstrodynamics.AstrodynamicalFrames.AstrodynamicsFrame — Type

abstract type AstrodynamicsFrame

A supertype for all coordinate frames used in space.

GeneralAstrodynamics.AstrodynamicalFrames.AstrodynamicsTransform — Type

abstract type AstrodynamicsTransform

A supertype for all astrodynamics coordinate frame transformations.

GeneralAstrodynamics.AstrodynamicalFrames.BarycentricInertial — Type

abstract type BarycentricInertial <: Inertial

A supertype for all barycentric-ineretial coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.BarycentricRotating — Type

abstract type BarycentricRotating <: Rotating

A supertype for all barycentric-rotating coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.BodycentricInertial — Type

abstract type BodycentricInertial <: Inertial

A supertype for all bodycentric-inertial coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.BodycentricRotating — Type

abstract type BodycentricRotating <: Rotating

A supertype for all bodycentric-rotating coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.Inertial — Type

abstract type Inertial <: AstrodynamicsFrame

A supertype for all inertial coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.Rotating — Type

abstract type Rotating <: AstrodynamicsFrame

A supertype for all rotating coordinate frames in space.

GeneralAstrodynamics.AstrodynamicalFrames.Transform — Type

struct Transform{F1<:AstrodynamicsFrame, F2<:AstrodynamicsFrame, T1<:CoordinateTransformations.Transformation, T2<:CoordinateTransformations.Transformation} <: AstrodynamicsTransform

A generic frame transformation between two AstrodynamicsFrame types. Converts F1 toF2.

transform_position::CoordinateTransformations.Transformation

transform_velocity::CoordinateTransformations.Transformation

GeneralAstrodynamics.AstrodynamicalFrames.Transform — Method

Transform(position_transform, velocity_transform, from, to)

An outer constructor for Transform instances. Provide position and velocity transformations, and the initial and final reference frames via function arguments. This is an alternate syntax for Transform{InitialFrame, FinalFrame}(position_transform, velocity_transform).

GeneralAstrodynamics.AstrodynamicalFrames.isbarycentric — Method

Returns true or false to indicate whether the provided frame is barycentric.

GeneralAstrodynamics.AstrodynamicalFrames.isbodycentric — Method

isbodycentric(_)

Returns true or false to indicate whether the provided frame is bodycentric.

GeneralAstrodynamics.AstrodynamicalFrames.isinertial — Method

Returns true or false to indicate whether the provided frame is Inertial.

GeneralAstrodynamics.AstrodynamicalFrames.isrotating — Method

Returns true or false to indicate whether the provided frame is Rotating.

GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit — Type

A supertype for all single-point orbit descriptions. Parameterized by coordinate frame, floating point type, mass unit, lenth unit, time unit, epoch type, state type, and parameter type (in order).

GeneralAstrodynamics.AstrodynamicalStates.CR3BPOrbit — Type

struct Orbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:CR3BPParameters} <: GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:CR3BPParameters}

An alias for Orbit instances about CR3BP systems.

GeneralAstrodynamics.AstrodynamicalStates.CR3BPParameters — Type

All parameters required for the Circular Restricted Three-body Problem.

GeneralAstrodynamics.AstrodynamicalStates.CartesianOrbit — Type

struct Orbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:GeneralAstrodynamics.AstrodynamicalStates.ParameterVector} <: GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:GeneralAstrodynamics.AstrodynamicalStates.ParameterVector}

An alias for Orbit instances about any systems with CartesianState descriptions.

GeneralAstrodynamics.AstrodynamicalStates.CartesianR2BPOrbit — Type

struct Orbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:R2BPParameters} <: GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:CartesianState, var"#s35"<:R2BPParameters}

An alias for Orbit instances about R2BP systems with CartesianState descriptions.

GeneralAstrodynamics.AstrodynamicalStates.CartesianState — Type

A Cartesian state vector with length 6. Internally uses MVector and LVector to store data. Data is accessible via labels, which are x, y, z, ẋ, ẏ, ż, r, v, where r access x,y,z and v accesses ẋ, ẏ, ż.

GeneralAstrodynamics.AstrodynamicalStates.CartesianState — Method

Constructs a CartesianState from provided position and velocity vectors.

GeneralAstrodynamics.AstrodynamicalStates.CartesianState — Method

Constructs a CartesianState.

GeneralAstrodynamics.AstrodynamicalStates.KeplerianR2BPOrbit — Type

struct Orbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:KeplerianState, var"#s35"<:R2BPParameters} <: GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:KeplerianState, var"#s35"<:R2BPParameters}

An alias for Orbit instances about R2BP systems with KeplerianState descriptions.

GeneralAstrodynamics.AstrodynamicalStates.KeplerianState — Type

A Keplerian state vector with length 6. Internally uses MVector and LVector to store data. Data is accessible via labels, which are e, a, i, Ω, ω, ν.

GeneralAstrodynamics.AstrodynamicalStates.KeplerianState — Method

Constructs a KeplerianState from provided position and velocity vectors.

GeneralAstrodynamics.AstrodynamicalStates.KeplerianState — Method

Constructs a KeplerianState from any AbstractVector.

GeneralAstrodynamics.AstrodynamicalStates.Orbit — Type

An orbit, described by a StateVector, with parameters described by a ParameterVector.

GeneralAstrodynamics.AstrodynamicalStates.Orbit — Type

Outer constructor for Orbits.

GeneralAstrodynamics.AstrodynamicalStates.ParameterVector — Type

A supertype for parameter representations in astrodynamics.

GeneralAstrodynamics.AstrodynamicalStates.ParameterizedLabelledArray — Type

abstract type ParameterizedLabelledArray{F, N, T, L} <: DenseArray{F, N}

A supertype for types that function like LabelledArray.LArray or LabelledArray.SLArray instances, but are under a new type tree. This is used in GeneralAstrodynamics for parameterizing astrodynamics state vectors by physical units.

Note

All subtypes must have only one field: a LabelledArrays.LArray or LabelledArrays.SLArray field called __rawdata. All methods on this abstract type require this field to be called __rawdata!

Nearly all code which acts on this type is copied and / or modified from LabelledArrays.jl source code. All credit goes to LabelledArrays.jl developers. The LabelledArrays.jl LICENSE file is provided in this docstring under Julia's Extended Help docstring section.

Extended help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

GeneralAstrodynamics.AstrodynamicalStates.R2BPOrbit — Type

struct Orbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:Union{CartesianState, KeplerianState}, var"#s35"<:R2BPParameters} <: GeneralAstrodynamics.AstrodynamicalStates.AbstractOrbit{FR, F, MU, LU, TU, AU, E, var"#s36"<:Union{CartesianState, KeplerianState}, var"#s35"<:R2BPParameters}

An alias for Orbit instances about R2BP systems.

GeneralAstrodynamics.AstrodynamicalStates.R2BPParameters — Type

All parameters required for the Restricted Two-body Problem.

GeneralAstrodynamics.AstrodynamicalStates.StateVector — Type

A supertype for all state representations in astrodynamics.

ArrayInterface.restructure — Method

restructure(x, y)

restructure(x, y)

Reshapes a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.convert — Method

convert(_, x)

convert(_, x)

Converts the underlying floating point type for a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.convert — Method

convert(_, params)

Converts between CR3BPParameters.

Base.convert — Method

convert(_, params)

Converts between R2BPParameters.

Base.convert — Method

convert(_, x)

convert(_, x)

Converts the underlying floating point type for a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.convert — Method

convert(_, x)

convert(_, x)

Converts the underlying floating point type for a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.copy — Method

copy(x)

copy(x)

Shallow copies a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.copyto! — Method

copyto!(x, y)

copyto!(x, y)

Copies one ParameterizedLabelledArray to another.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.dataids — Method

dataids(A)

dataids(A)

Implements dataids for a ParameterizedLabelledArray instance.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.deepcopy — Method

deepcopy(x)

deepcopy(x)

Deep copies a ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.elsize — Method

Returns the memory stride for any ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.getindex — Method

getindex(state, args)

getindex(state, args)

Returns the index of the ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.getproperty — Method

getproperty(x, s)

getproperty(x, s)

Overrides Base.getproperty for all ParameterizedLabelledArray instances.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.length — Method

length(_)

The length of any StateVector is 6!

Base.length — Method

Returns the length of the parameter vector.

Base.propertynames — Method

propertynames(_)

propertynames(_)

Returns dot-accessible property names for a ParameterizedLabelledArray.

Base.setindex! — Method

setindex!(state, args)

setindex!(state, args)

Sets the index of the ParameterizedLabelledArray.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.setproperty! — Method

setproperty!(x, s, y)

setproperty!(x, s, y)

Sets indices of a ParameterizedLabelledArray via label.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.show — Method

show(io, state; showfloats, space)

Displays CR3BPParameters.

Base.show — Method

show(io, state; showfloats, space)

Displays a CartesianState.

Base.show — Method

show(io, state; showfloats, space)

Displays a KeplerianState.

Base.show — Method

show(io, , state; showfloats, space)

Displays a CartesianState.

Base.show — Method

show(io, , state; showfloats, space)

Displays a KeplerianState.

Base.show — Method

show(io, state; showfloats, space)

Displays R2BPParameters.

Base.similar — Method

similar(x, _, dims)

similar(x, _, dims)

Overrides similar for ParameterizedLabelledArray instances.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

Base.size — Method

size(_)

The size of any StateVector is (6,)!

Base.size — Method

Returns the size of the parameter vector.

Base.unsafe_convert — Method

unsafe_convert(_, a)

unsafe_convert(_, a)

Provides unsafe_convert for ParameterizedLabelledArray types for use with LAPACK.

This source code which provides this functionality was copied directly from LabelledArrays.jl source code. The LabelledArrays.jl license text is provided in Julia's Extended Help section (accessible via @doc, or ?? in Julia's REPL).

Extended Help

LabelledArrays.jl License

The LabelledArrays.jl package is licensed under the MIT "Expat" License:

Copyright (c) 2017: Christopher Rackauckas.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

GeneralAstrodynamics.AstrodynamicalStates.angularunit — Method

Returns the angular unit of the parameter vector.

GeneralAstrodynamics.AstrodynamicalStates.angularunit — Method

Returns the angularunit of the state vector.

GeneralAstrodynamics.AstrodynamicalStates.epoch — Method

epoch(orbit)

Returns the epoch (timestamp) for the Orbit.

GeneralAstrodynamics.AstrodynamicalStates.get_a — Method

get_a(state)

Returns a.

GeneralAstrodynamics.AstrodynamicalStates.get_e — Method

get_e(state)

Returns e.

GeneralAstrodynamics.AstrodynamicalStates.get_i — Method

get_i(state)

Returns i.

GeneralAstrodynamics.AstrodynamicalStates.get_r — Method

get_r(state)

Returns r.

GeneralAstrodynamics.AstrodynamicalStates.get_v — Method

get_v(state)

Returns v.

GeneralAstrodynamics.AstrodynamicalStates.get_x — Method

get_x(state)

Returns x.

GeneralAstrodynamics.AstrodynamicalStates.get_y — Method

get_y(state)

Returns y.

GeneralAstrodynamics.AstrodynamicalStates.get_z — Method

get_z(state)

Returns z.

GeneralAstrodynamics.AstrodynamicalStates.get_ż — Method

get_ż(state)

Returns ż.

GeneralAstrodynamics.AstrodynamicalStates.get_Ω — Method

get_Ω(state)

Returns Ω.

GeneralAstrodynamics.AstrodynamicalStates.get_ν — Method

get_ν(state)

Returns ν.

GeneralAstrodynamics.AstrodynamicalStates.get_ω — Method

get_ω(state)

Returns `ω.

GeneralAstrodynamics.AstrodynamicalStates.get_ẋ — Method

get_ẋ(state)

Returns ẋ.

GeneralAstrodynamics.AstrodynamicalStates.get_ẏ — Method

get_ẏ(state)

Returns ẏ.

GeneralAstrodynamics.AstrodynamicalStates.lengthunit — Method

Returns the length unit of the parameter vector.

GeneralAstrodynamics.AstrodynamicalStates.lengthunit — Method

Returns the lengthunit of the state vector.

GeneralAstrodynamics.AstrodynamicalStates.massparameter — Method

massparameter(system)

Returns the mass parameter of the R2BP system.

GeneralAstrodynamics.AstrodynamicalStates.massparameters — Method

massparameters(system)

Returns the mass parameters of the CR3BP system.

GeneralAstrodynamics.AstrodynamicalStates.massparamunit — Method

Returns the mass-parameter unit of the state vector.

GeneralAstrodynamics.AstrodynamicalStates.massunit — Method

Returns the mass unit of the parameter vector.

GeneralAstrodynamics.AstrodynamicalStates.name — Method

Returns the name of the parameter vector.

GeneralAstrodynamics.AstrodynamicalStates.normalized_massparameter — Method

normalized_massparameter(system)

Returns the mass parameter of the CR3BP system.

GeneralAstrodynamics.AstrodynamicalStates.primary_massparameter — Method

primary_massparameter(system)

Returns the primary mass parameter of the CR3BP system.

GeneralAstrodynamics.AstrodynamicalStates.secondary_massparameter — Method

secondary_massparameter(system)

Returns the secondary mass parameter of the CR3BP system.

GeneralAstrodynamics.AstrodynamicalStates.state — Method

state(orbit)

Returns the state vector for the Orbit.

GeneralAstrodynamics.AstrodynamicalStates.statevector — Method

statevector(state)

Returns the whole state vector, with units.

GeneralAstrodynamics.AstrodynamicalStates.statevector — Method

statevector(state)

Returns the whole state vector, with units.

GeneralAstrodynamics.AstrodynamicalStates.system — Method

system(orbit)

Returns the parameter vector for the Orbit.

GeneralAstrodynamics.AstrodynamicalStates.timeunit — Method

Returns the time unit of the parameter vector.

GeneralAstrodynamics.AstrodynamicalStates.timeunit — Method

Returns the timeunit of the state vector.

GeneralAstrodynamics.AstrodynamicalStates.velocityunit — Method

Returns the velocity unit of the state vector.

GeneralAstrodynamics.AstrodynamicalStates.velocityunit — Method

Returns the angularunit of the state vector.

GeneralAstrodynamics.AstrodynamicalCalculations.Circular — Type

abstract type Circular <: GeneralAstrodynamics.AstrodynamicalCalculations.ConicSection

An abstract type representing the Circular conic section.

GeneralAstrodynamics.AstrodynamicalCalculations.ConicSection — Type

abstract type ConicSection

An abstract type for all conic sections.

GeneralAstrodynamics.AstrodynamicalCalculations.Elliptical — Type

abstract type Elliptical <: GeneralAstrodynamics.AstrodynamicalCalculations.ConicSection

An abstract type representing the Elliptical conic section.

GeneralAstrodynamics.AstrodynamicalCalculations.Hyperbolic — Type

abstract type Hyperbolic <: GeneralAstrodynamics.AstrodynamicalCalculations.ConicSection

An abstract type representing the Hyperbolic conic section.

GeneralAstrodynamics.AstrodynamicalCalculations.Parabolic — Type

abstract type Parabolic <: GeneralAstrodynamics.AstrodynamicalCalculations.ConicSection

An abstract type representing the Parabolic conic section.

GeneralAstrodynamics.AstrodynamicalCalculations.C3 — Method

C3(r, v, μ)

Returns C3 value.

GeneralAstrodynamics.AstrodynamicalCalculations.SOA — Method

SOA(a, m, M)

Sphere of activity.

GeneralAstrodynamics.AstrodynamicalCalculations.SOI — Method

SOI(a, m, M)

Sphere of influence.

GeneralAstrodynamics.AstrodynamicalCalculations.apoapsis_radius — Method

apoapsis_radius(a, e)

Returns apoapsis distance, rₐ.

GeneralAstrodynamics.AstrodynamicalCalculations.cartesian — Method

cartesian(e, a, i, Ω, ω, ν, μ)

Returns a Cartesian representation of a Keplerian two-body orbital state in an inertial frame, centered at the center of mass of the central body. Algorithm taught in ENAE601.

GeneralAstrodynamics.AstrodynamicalCalculations.conic — Method

conic(e)

Returns the conic section, as specified by eccentricity e.

GeneralAstrodynamics.AstrodynamicalCalculations.distance — Method

distance(p, e, ν)

Returns distance, r.

GeneralAstrodynamics.AstrodynamicalCalculations.eccentricity — Method

eccentricity(rᵢ, vᵢ, μ)

Returns orbital eccentricity, e.

GeneralAstrodynamics.AstrodynamicalCalculations.eccentricity_vector — Method

eccentricity_vector(rᵢ, vᵢ, μ)

Returns orbital eccentricity vector e̅.

GeneralAstrodynamics.AstrodynamicalCalculations.hohmann — Method

hohmann(r₁, r₂, μ)

Computes a Hohmann transfer, and returns the departure and arrival velocity vectors.

GeneralAstrodynamics.AstrodynamicalCalculations.keplerian — Method

keplerian(rᵢ, vᵢ, μ)

Returns a Keplarian representation of a Cartesian orbital state. Algorithm taught in ENAE601.

GeneralAstrodynamics.AstrodynamicalCalculations.mean_motion — Method

mean_motion(a, μ)

Returns mean motion, n.

GeneralAstrodynamics.AstrodynamicalCalculations.periapsis_radius — Method

periapsis_radius(a, e)

Returns periapsis distance, rₚ.

GeneralAstrodynamics.AstrodynamicalCalculations.perifocal — Method

perifocal(a, e, ν, μ)

Returns position and velocity vectors in the Perifocal frame.

GeneralAstrodynamics.AstrodynamicalCalculations.period — Method

period(a, μ)

Returns the orbital period.

GeneralAstrodynamics.AstrodynamicalCalculations.semi_parameter — Method

semi_parameter(a, e)

Returns semilatus parameter, p.

GeneralAstrodynamics.AstrodynamicalCalculations.semimajor_axis — Method

semimajor_axis(r, v, μ)

Returns semimajor axis parameter, a.

GeneralAstrodynamics.AstrodynamicalCalculations.spatial — Method

spatial(i, Ω, ω, rₚ, vₚ)

Returns a spatial representation of the provied Perifocal state.

GeneralAstrodynamics.AstrodynamicalCalculations.specific_angular_momentum — Method

specific_angular_momentum(rᵢ, vᵢ)

Returns scalar specific angular momentum vector, h.

GeneralAstrodynamics.AstrodynamicalCalculations.specific_angular_momentum_vector — Method

specific_angular_momentum_vector(rᵢ, vᵢ)

Returns specific angular momentum vector, h̅.

GeneralAstrodynamics.AstrodynamicalCalculations.specific_energy — Method

specific_energy(a, μ)

Returns specific orbital energy, ϵ.

GeneralAstrodynamics.AstrodynamicalCalculations.specific_potential_energy — Method

specific_potential_energy(r, μ)

Returns potential energy for an orbit about a RestrictedTwoBodySystem.

GeneralAstrodynamics.AstrodynamicalCalculations.speed — Method

speed(p, e, ν, a, μ)

Returns the instantaneous velocity, v, for any orbital representation.

GeneralAstrodynamics.AstrodynamicalCalculations.speed — Method

speed(r, a, μ)

Returns instantaneous velocity, v, for any orbital representation.

GeneralAstrodynamics.AstrodynamicalCalculations.time_since_periapsis — Method

time_since_periapsis(n, e, E)

Returns time since periapsis, t.

GeneralAstrodynamics.AstrodynamicalCalculations.true_anomoly — Method

true_anomoly(r, h, e, μ)

Returns true anomoly, ν.

GeneralAstrodynamics.AstrodynamicalCalculations.v_infinity — Method

v_infinity(r, v, μ)

Returns v∞.