Reference

All exported names.

AstrodynamicalSolvers.AstrodynamicalSolvers — Module

Provides astrodynamical solvers, including Lyapunov and halo orbit correctors.

Extended help

License

MIT License

Copyright (c) 2023 Joseph D Carpinelli

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Exports

CR3BSolvers

Propagation

convergent_manifold

divergent_manifold

halo

lyapunov

monodromy

propagate

propagate!

Imports

AstrodynamicalSolvers.CR3BSolvers

AstrodynamicalSolvers.Propagation

Base

Core

DocStringExtensions

Reexport

AstrodynamicalSolvers.Propagation — Module

Wrappers around SciML differential equation solvers for fast and convenient orbit propagation.

Extended Help

Exports

convergent_manifold

divergent_manifold

monodromy

propagate

propagate!

Imports

AstrodynamicalCalculations

AstrodynamicalModels

Base

Core

DocStringExtensions

ModelingToolkit

OrdinaryDiffEqVerner

SciMLBase

StaticArrays

AstrodynamicalSolvers.Propagation.convergent_manifold — Method

convergent_manifold(u, μ, Δt; eps, trajectories, kwargs...)

Return a vector of orbits along the manifold which converges to the provided halo orbit.

AstrodynamicalSolvers.Propagation.divergent_manifold — Method

divergent_manifold(u, μ, Δt; eps, trajectories, kwargs...)

Return a vector of orbits along the manifold which diverges from the provided halo orbit.

AstrodynamicalSolvers.Propagation.monodromy — Method

monodromy(
    u,
    μ,
    T,
    f;
    algorithm,
    reltol,
    abstol,
    save_everystep,
    kwargs...
)

Solve for the monodromy matrix of the periodic orbit.

AstrodynamicalSolvers.Propagation.monodromy — Method

monodromy(orbit, Δt; algorithm, reltol, abstol, kwargs...)

Compute the monodromy matrix for any periodic orbit.

AstrodynamicalSolvers.Propagation.propagate! — Method

propagate!(
    orbit,
    Δt;
    stm,
    algorithm,
    reltol,
    abstol,
    kwargs...
)

Numerically integrate the orbit forward (or backward) in time, modifying the state vector in-place within the AstrodynamicalOrbit instance.

AstrodynamicalSolvers.Propagation.propagate — Method

propagate(
    orbit,
    Δt;
    stm,
    algorithm,
    reltol,
    abstol,
    kwargs...
)

Numerically integrate the orbit forward (or backward) in time, and return a new AstrodynamicalOrbit instance with identical parameters to the provided orbit.